// aws/shopify/webhooks/ordersCancelled/types/admin.generated.d.ts
export type FulfillmentOrderQuery = { fulfillmentOrder?: AdminTypes.Maybe<Pick<AdminTypes.FulfillmentOrder, 'orderId'>> };

export type OrderLineItemsQueryVariables = AdminTypes.Exact<{
  id: AdminTypes.Scalars['ID']['input'];
}>;


export type OrderLineItemsQuery = { order?: AdminTypes.Maybe<{ lineItems: { edges: Array<{ node: (
          Pick<AdminTypes.LineItem, 'id'>
          & { variant?: AdminTypes.Maybe<Pick<AdminTypes.ProductVariant, 'id'>> }
        ) }>, pageInfo: Pick<AdminTypes.PageInfo, 'hasNextPage' | 'endCursor'> } }> };

export type SubsequentOrderLineItemsQueryVariables = AdminTypes.Exact<{
  id: AdminTypes.Scalars['ID']['input'];
  after: AdminTypes.Scalars['String']['input'];
}>;


export type SubsequentOrderLineItemsQuery = { order?: AdminTypes.Maybe<{ lineItems: { edges: Array<{ node: (
          Pick<AdminTypes.LineItem, 'id'>
          & { variant?: AdminTypes.Maybe<Pick<AdminTypes.ProductVariant, 'id'>> }
        ) }>, pageInfo: Pick<AdminTypes.PageInfo, 'hasNextPage' | 'endCursor'> } }> };

export type MMutationVariables = AdminTypes.Exact<{
  input: AdminTypes.RefundInput;
}>;


export type MMutation = { refundCreate?: AdminTypes.Maybe<{ userErrors: Array<Pick<AdminTypes.UserError, 'field' | 'message'>>, refund?: AdminTypes.Maybe<(
      Pick<AdminTypes.Refund, 'id' | 'note'>
      & { totalRefundedSet: { presentmentMoney: Pick<AdminTypes.MoneyV2, 'amount'> } }
    )> }> };